refactor back navigation into separate hook
