+++ src/hooks/use-back-navigate.ts
@@ -1,0 +1,15 @@
+import { useNavigate } from "@tanstack/react-router"
+
+export const useBackNavigate = (path: string = '..'): () => Promise<void> => {
+  const navigate = useNavigate()
+
+  return async () => {
+    if (window.history.length > 1) {
+      window.history.back()
+
+      return
+    }
+
+    await navigate({ to: path })
+  }
+}
